Clean up use of 'UNSET' string
In Puppet, the `undef` keyword evaluates to boolean false. Thus, allows
for more concise code when evaluating to see if a variable exists or
not.

This commit will use the `undef` keyword in place of the string 'UNSET'.
It also prefers simple conditional expressions over string comparisons
in both the `::redis` class and the erb templates.