Use Redis bind default as parameter default value
By default, Redis will attempt to listen on all interfaces available on
a node. However, the `::redis` class currently uses restrictive default
parameters that do not operate as a default Redis installation is
expected to.

This commit uses 0.0.0.0 as the default bind address to replicate the
default equivalent of Redis. It also no long allows the bind directive to
be absent since the default behavior is equivalent to removing the bind
directive.

Generally speaking, this is a better best practice for module design. It
imposes less policy on the people who will be using the module and
leaves it as an exercise for them at implementation time.