Adding support for Redis Sentinel
This is implemented in the similar way as redis-server, supporting
same feature as the other and also being able to configure Sentinel
with every parameter.

Since there is not a Sentinel package, this does not install Sentinel,
but expect to have the init scripts to manage the service.

Also, a special thing about this implementation is that the sentinel
config file is not overriden in puppet executions, this is due to
Sentinel rewrites this file everytime the is a failover process.