Break out configuration files based on the version
The configuration file for redis server is ~2800 LOC and it quite
difficult to eyeball it without getting confused as you scroll through
the file. This changes that so that each version is in it's own file
based on the `$redis_version_real` value provided by the module. The
only change made to the `$redis_version_real` is that all periods are
being replaced by hyphens for a little more readablilty.